requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.2
cycler==0.10.0
Django==3.0.6
django-picklefield==2.1.1
djangorestframework==3.11.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
opencv-contrib-python==4.2.0.34
pickle-mixin==1.0.2
Pillow==7.1.2
psutil==5.7.0
pylint==2.5.2
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyqt5-tools==5.14.2.1.7b3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wrapt==1.12.1
